add comment and reorg code
